Split services/queries.py into 2 files
services/metadata_queries.py and services/vcf_queries.py

The logic was already quite separated. The file was long and the split
hopefully makes it easier to navigate and lower cognitive load when
working on these files.
