+++ docs/development/worker_metrics.md
@@ -48,7 +48,7 @@ Note 2: At the time of writing, the `bcftools` subprocesses in DivBase are imple
 
 Resource specifications in k8s is enforced per container (not per pod, the pod resouce usage is the sum of all its containers). CPU is measured in cores: 1 = one full core, 500m (millicore) is half a core, etc. Important to know is that when a container exceedes its CPU limitations in k8s, it is throttled but not killed; exceeding memory limitations result in containers being kill, but more on that in the memory section below.
 
-In the current DivBase implementation, bcftools subprocesses run sequentially and not in parallel. Each subprocess completes before the next one starts (see `run_current_command()` in `packages/divbase-api/src/divbase_api/services/queries.py`). This means that **allocating more CPU cores will NOT reduce wall time** in the current implementation.
+In the current DivBase implementation, bcftools subprocesses run sequentially and not in parallel. Each subprocess completes before the next one starts (see `run_current_command()` in `packages/divbase-api/src/divbase_api/services/vcf_queries.py`). This means that **allocating more CPU cores will NOT reduce wall time** in the current implementation.
 
 CPU limits in k8s still matter for preventing throttling, but won't speed up tasks unless parallelization is implemented. To configure this for the k8s deployment use `requests` for the minimum guaranteed CPU resources and `limits` for the maximum allowed CPU resources to avoid throttling.
 
@@ -100,7 +100,7 @@ The metrics are captured in the tasks themselves in `packages/divbase-api/src/di
 
 Metrics for wall time is captured with `time.time()`, and CPU time, average and peak RSS memory is captured with `psutil`. The Python process and the `bcftools` subprocesses are monitored separately. The Python process monitoring is initiated with `process = psutil.Process()`; CPU monitoring is started with `process.cpu_times()`; memory usage is collected by sampling RSS over the course of the process with `process.memory_info().rss` as explained in the [Memory Usage](#memory-usage-not-additive-across-processes) section above and this is implemented as a class named `MemoryMonitor` in `packages/divbase-api/src/divbase_api/worker/metrics.py`.
 
-The `bcftools` subprocesses CPU and memory monitoring is found in `packages/divbase-api/src/divbase_api/services/queries.py`. When the code spawns a subprocess, it uses `subprocess.Popen` (instead of the commonly used `subprocess.run`). The PID of each spawned process is collected and `psutil.Process(proc.pid)` is used to collect the data from each PID with the `psutil` methods `.cpu_times()` and `.memory_info().rss` just like for the Python process.
+The `bcftools` subprocesses CPU and memory monitoring is found in `packages/divbase-api/src/divbase_api/services/vcf_queries.py`. When the code spawns a subprocess, it uses `subprocess.Popen` (instead of the commonly used `subprocess.run`). The PID of each spawned process is collected and `psutil.Process(proc.pid)` is used to collect the data from each PID with the `psutil` methods `.cpu_times()` and `.memory_info().rss` just like for the Python process.
 
 After a task completes (regardless of Celery task status, e.g. SUCCESS or FAILED), the metrics are stored in an in-memory cache structured as a dictionary of dictionaries. This is done with the helper function `_record_task_metrics` in `packages/divbase-api/src/divbase_api/worker/tasks.py`, which calls `store_task_metric()` to save each metric value along with a timestamp. The cached values are then used to update the respective Prometheus gauges via `update_prometheus_gauges_from_cache()` in `packages/divbase-api/src/divbase_api/worker/metrics.py`. The metrics are stored per task using the DivBase job ID (not the Celery task UUID!) and task name as labels. The metrics are exposed via the Prometheus metrics endpoint at `/metrics`.
