+++ tests/e2e_integration/queries_integration/conftest.py
@@ -8,7 +8,8 @@ import pandas as pd
 import pytest
 from celery import current_app
 
-from divbase_api.services.queries import BCFToolsInput, BcftoolsQueryManager, SampleFileMapping, SidecarQueryManager
+from divbase_api.services.metadata_queries import SidecarQueryManager
+from divbase_api.services.vcf_queries import BCFToolsInput, BcftoolsQueryManager, SampleFileMapping
 
 
 @pytest.fixture
